dags/1_first_dag.py
from airflow.sdk import dag, task

@dag(
        dag_id= "first_dag"
)
def first_dag():

    @task.python
    def first_task():
        print("This is the first task")

    @task.python
    def second_task():
        print("This is the second task")

    
    @task.python
    def third_task():
        print("This is the third task")




    # Defining the task dependencies
    first_task() >> second_task() >> third_task()


# Registering the DAG
first_dag = first_dag()